[BACKLOG-2875] Refactored for compatibility with gwt and data-access build
